Revert accidental test classes names change
