+++ tests/Ackintosh/Ganesha/Storage/Adapter/PredisRedisTest.php
@@ -5,7 +5,7 @@ namespace Ackintosh\Ganesha\Storage\Adapter;
 use Predis\Client;
 use Predis\ClientInterface;
 
-class PredisRedisSpec extends AbstractRedisSpec
+class PredisRedisTest extends AbstractRedisSpec
 {
     protected function getRedisConnection(): ClientInterface
     {
